+++ packages/site/src/__test__/utils/coreColorTokens.test.js
@@ -1,7 +1,7 @@
 import {
 	getCoreColorsFromTokens,
 	getUpdatedTokensWithCoreColors,
-} from "../helpers";
+} from "../../utils/coreColorTokens";
 
 describe("Core Color Tokens", () => {
 	const tokensBase = {
